requirements.txt
# Browser automation dependencies
browser-use>=0.1.0
langchain-openai>=0.1.0
playwright>=1.40.0

# Additional dependencies that might be needed
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
allure-pytest==2.14.3
pytest-asyncio==1.0.0
pytest>=7.0.0